Move collection and iterator packages to collections module
