+++ apso-collections/src/main/scala/com/velocidi/apso/collection/TypedMap.scala
