+++ apso-collections/src/main/scala/com/velocidi/apso/iterator/MergedBufferedIterator.scala
