+++ apso-collections/src/test/scala/com/velocidi/apso/collection/TrieSpec.scala
