+++ apso-collections/src/test/scala/com/velocidi/apso/iterator/CircularIteratorSpec.scala
