+++ core/src/test/scala/com/velocidi/apso/ImplicitsSpec.scala
@@ -9,103 +9,6 @@ import com.velocidi.apso.Implicits._
 
 class ImplicitsSpec extends Specification with ScalaCheck with FutureExtraMatchers {
 
-  "An ApsoSeq" should {
-
-    "merge two sorted collections" in {
-      List(1, 5, 6).mergeSorted[Int, List[Int]](List.empty[Int]) === List(1, 5, 6)
-      List.empty[Int].mergeSorted(List(2, 4)) === List(2, 4)
-      Seq(2).mergeSorted(Seq(5)) === List(2, 5)
-      Seq(5).mergeSorted(Seq(2)) === List(2, 5)
-
-      List(1, 3, 5).mergeSorted(Stream(2, 4)) === List(1, 2, 3, 4, 5)
-      List(1, 3, 5).mergeSorted(Stream(2, 4)) must beAnInstanceOf[List[_]]
-
-      Stream(1, 3, 5).mergeSorted(List(2, 4)) === Stream(1, 2, 3, 4, 5)
-      Stream(1, 3, 5).mergeSorted(List(2, 4)) must beAnInstanceOf[Stream[_]]
-
-      trait Base { val x: Int }
-      case class Impl1(x: Int) extends Base
-      case class Impl2(x: Int) extends Base
-
-      implicit def ord[T <: Base] = new Ordering[T] {
-        override def compare(a: T, b: T): Int = a.x - b.x
-      }
-
-      Seq(Impl1(3)).mergeSorted(Seq(Impl2(5))) === Seq(Impl1(3), Impl2(5))
-      Seq(Impl1(3)).mergeSorted(Seq(Impl2(5))) must beAnInstanceOf[Seq[Base]]
-    }
-  }
-
-  "An ApsoTraversableOnce" should {
-
-    "calculate correctly the average of a list of integral values" in {
-      List[Byte](1, 4, 3, 2, 5).average === 3
-      List[Short](2, 8, 6, 4, 10).average === 6
-      List[Int](3, 12, 9, 6, 15).average === 9
-      List[Long](4, 16, 12, 8, 20).average === 12
-      List[Byte](1, 2).average === 1
-      List[Int](1, 2).average === 1
-      List[Short](1, 2).average === 1
-      List[Long](1, 2).average === 1
-    }
-
-    "calculate correctly the average of a list of fractional values" in {
-      List[Float](1, 4, 3, 2, 5).average === 3.0
-      List[Double](2, 8, 6, 4, 10).average === 6.0
-      List[Float](1, 2).average === 1.5
-      List[Double](1, 2).average === 1.5
-    }
-
-    "throw an exception when the average is called on an empty traversable" in {
-      List[Int]().average must throwA[IllegalArgumentException]
-    }
-  }
-
-  "An ApsoBufferedIterator" should {
-
-    "merge sorted iterators" in {
-      List(1, 5, 6).iterator.buffered.mergeSorted(List.empty[Int].iterator.buffered).toList === List(1, 5, 6)
-      List.empty[Int].iterator.buffered.mergeSorted(List(2, 4).iterator.buffered).toList === List(2, 4)
-      Seq(2).iterator.buffered.mergeSorted(Seq(5).iterator.buffered).toList === List(2, 5)
-      Seq(5).iterator.buffered.mergeSorted(Seq(2).iterator.buffered).toList === List(2, 5)
-
-      List(1, 3, 5).iterator.buffered.mergeSorted(Stream(2, 4).iterator.buffered).toList === List(1, 2, 3, 4, 5)
-
-      Stream(1, 3, 5).iterator.buffered.mergeSorted(List(2, 4).iterator.buffered).toStream === Stream(1, 2, 3, 4, 5)
-
-      trait Base { val x: Int }
-      case class Impl1(x: Int) extends Base
-      case class Impl2(x: Int) extends Base
-
-      implicit def ord[T <: Base] = new Ordering[T] {
-        override def compare(a: T, b: T): Int = a.x - b.x
-      }
-
-      Seq(Impl1(3)).iterator.buffered.mergeSorted(Seq(Impl2(5)).iterator.buffered).toSeq === Seq(Impl1(3), Impl2(5))
-      Seq(Impl1(3)).iterator.buffered.mergeSorted(Seq(Impl2(5)).iterator.buffered).toSeq must beAnInstanceOf[Seq[Base]]
-
-      List(1, 4, 7).iterator.buffered.mergeSorted(List(2, 5, 8).iterator.buffered).mergeSorted(List(3, 6, 9).iterator.buffered).toList ===
-        List(1, 2, 3, 4, 5, 6, 7, 8, 9)
-    }
-  }
-
-  "An ApsoMap" should {
-
-    "support the twoWayMerge method" in {
-      val m1 = Map(1 -> 1, 2 -> 2, 3 -> 3)
-      val m2 = Map(3 -> 3, 4 -> 4, 5 -> 5)
-
-      m1.twoWayMerge(m2)(_ + _) ===
-        Map(1 -> 1, 2 -> 2, 3 -> 6, 4 -> 4, 5 -> 5)
-    }
-
-    "support the mapKeys method" in {
-      val m = Map(1 -> 2, 3 -> 4, 5 -> 6, 7 -> 8)
-      m.mapKeys(_ * 2) ===
-        Map(2 -> 2, 6 -> 4, 10 -> 6, 14 -> 8)
-    }
-  }
-
   "An ApsoRandom" should {
 
     class MockRandom(elems: Double*) extends Random {
